refactor(#184): migrate init composer handlers to sdk/src/handlers/init
